+++ lib/errbit_github_plugin.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "errbit_github_plugin/version"
-require "errbit_github_plugin/error"
+require "errbit_github_plugin/authentication_error"
 require "errbit_github_plugin/issue_tracker"
 
 module ErrbitGithubPlugin
