+++ lib/errbit_github_plugin/authentication_error.rb
